Move to a single PopOver type with a hidden modal field to chose renderer
